Exact match item count and size are uint8.

+++ source/include/protocol/cs_ExactMatchFilterStructs.h
@@ -15,7 +15,7 @@ typedef const uint8_t* filter_entry_t;
  * Data content of the exact match filter.
  */
 struct __attribute__((__packed__)) exact_match_filter_data_t {
-	uint16_t itemCount;
-	uint16_t itemSize;
+	uint8_t itemCount;
+	uint8_t itemSize;
 	uint8_t itemArray[];
 };
